+++ crates/engine/src/lib.rs
@@ -23,6 +23,7 @@ pub mod protocol;
 pub mod query;
 pub mod render;
 pub mod render_channel;
+pub mod render_event;
 mod resource;
 pub mod schedule;
 pub mod system_param;
@@ -59,6 +60,7 @@ pub use query::{
 };
 pub use render::{MaterialHandle, MeshHandle, ObjectType, ParentEntity, Transform, Visibility};
 pub use render_channel::{ChannelRegistration, ExtractToFloats, RenderChannelRegistry};
+pub use render_event::{FrameEvent, RenderEvent, RenderEventRegistry};
 pub use schedule::Schedule;
 pub use system_param::{Access, Query, QueryMut, Res, ResMut, SystemParam};
 pub use virtual_time::VirtualTime;